data/requirements.txt
flask
pandas
numpy
matplotlib
scikit-optimize
torch
sklearn.preprocessing
keras
tensorflow
keras-tuner
keras-preprocessing